bff/pyproject.toml
[project]
name = "bff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "cryptography>=45.0.4",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 120
